% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_pheno_data.R
\name{calculate_time_since_baseline}
\alias{calculate_time_since_baseline}
\title{Calculate time since baseline}
\usage{
calculate_time_since_baseline(
  pheno,
  col.id.sample,
  col.id.subject,
  col.date,
  col.visit,
  col.time.diff
)
}
\arguments{
\item{pheno}{[data.frame] harmonized phenotype data}

\item{col.id.sample}{[vector(1)] column name or number with sample identifier}

\item{col.id.subject}{[vector(1)] column name or number with subject
identifier}

\item{col.date}{[vector(1)] column name or number with date of sample
collection (needs to be of class Date)}

\item{col.visit}{[vector(1)] column name or number with visit number (needs
to be numeric)}

\item{col.time.diff}{[character(1)] column name for new variable with time
since baseline}
}
\description{
If visit dates are available time since baseline visits (in days) is
calculated.
}
