% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_expr_data.R
\name{check_annotation_package}
\alias{check_annotation_package}
\title{Extract BrainArray annotation package from CEL file}
\usage{
check_annotation_package(cel.file, version = "25.0.0", temp.dir = tempdir())
}
\description{
Extract relevant BrainArray annotation package for mapping probes to Ensembl
genes based on header of CEL file. Install package if necessary. Code based
on SCAN.UPC::InstallBrainArrayPackage.
}
\keyword{internal}
