% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_pheno_data.R
\name{convert_numeric_variable}
\alias{convert_numeric_variable}
\title{Convert variable to numeric}
\usage{
convert_numeric_variable(x)
}
\description{
Extract numbers if necessary and convert to numeric.
}
\keyword{internal}
