% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_pheno_data.R
\name{convert_time_variable}
\alias{convert_time_variable}
\title{Convert variable to time in days}
\usage{
convert_time_variable(x, values)
}
\description{
Extract numbers if necessary, convert to numeric and transform to days.
}
\keyword{internal}
