% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_batch.R
\name{define_batches}
\alias{define_batches}
\title{Define batches based on scan date}
\usage{
define_batches(se, col.scan.date = "scan.date", diff.ignore = 1)
}
\arguments{
\item{se}{[\code{\link[SummarizedExperiment]{SummarizedExperiment-class}} object]
study data set}

\item{col.scan.date}{[character(1)] Column in colData() with scan
dates (default: scan.date as generated by the function
\code{\link{extract_scan_date}}}

\item{diff.ignore}{[numeric(1)] Time difference (in days) defined as
negligible (default: 1)}
}
\value{
\code{\link[SummarizedExperiment]{SummarizedExperiment-class}}
object with batch information added as column batch in colData()
}
\description{
Groups samples into batches based on scan date. Samples run within a short
time interval can be defined to belong to the same batch.
}
