% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_expr_data.R
\name{download_cel_file}
\alias{download_cel_file}
\title{Download CEL file (Affy) from ArrayExpress or GEO}
\usage{
download_cel_file(sample, temp.dir = tempdir())
}
\description{
Download CEL file (Affy) from ArrayExpress or GEO
}
\keyword{internal}
