% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_arrayexpress.R
\name{download_cel_file_AE}
\alias{download_cel_file_AE}
\title{Download CEL file (Affy) from ArrayExpress}
\usage{
download_cel_file_AE(sample, temp.dir = tempdir())
}
\description{
Download CEL file (Affy) from ArrayExpress
}
\keyword{internal}
