% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_geo.R
\name{download_cel_file_GEO}
\alias{download_cel_file_GEO}
\title{Download CEL file (Affy) from GEO}
\usage{
download_cel_file_GEO(sample, temp.dir = tempdir())
}
\description{
Download CEL file (Affy) from GEO
}
\keyword{internal}
