% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_arrayexpress.R
\name{download_files_AE}
\alias{download_files_AE}
\title{Download files from ArrayExpress}
\usage{
download_files_AE(ae.id, temp.dir = tempdir())
}
\description{
Download files from ArrayExpress and save in temporary directory. If files
are already available, they will not be downloaded again.
}
\keyword{internal}
