% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_pheno_data.R
\name{extract_identifier}
\alias{extract_identifier}
\title{Extract sample and subject identifiers}
\usage{
extract_identifier(pheno, col.id, ind.use.id = NULL)
}
\description{
Extract sample and subject identifiers
}
\keyword{internal}
