% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_arrayexpress.R
\name{extract_info_from_idf}
\alias{extract_info_from_idf}
\title{Extract information from IDF}
\usage{
extract_info_from_idf(idf, label)
}
\description{
Extract specific information from IDF file
}
\keyword{internal}
