% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_meta_data.R
\name{extract_meta_data}
\alias{extract_meta_data}
\title{Extract meta data}
\usage{
extract_meta_data(study.id, temp.dir = tempdir(), platform = NULL)
}
\arguments{
\item{study.id}{[character(1)] study identifier (GSEXXX for GEO, E-MTAB-XXX
for ArrayExpress)}

\item{temp.dir}{[character(1)] directory where temporary files should be
stored}

\item{platform}{[character(1)] platform to be selected in GEO
studies which use several platforms}
}
\value{
[list] meta data (content depending on information available at GEO
or ArrayExpress)
}
\description{
Extract meta data from GEO or ArrayExpress depending on study
identifier.
}
\details{
Different meta data information is available at GEO and ArrayExpress.
}
\examples{
meta.data.l = extract_meta_data(study.id = "GSE67785")
}
