% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_arrayexpress.R
\name{extract_meta_data_AE}
\alias{extract_meta_data_AE}
\title{Extract meta data from ArrayExpress}
\usage{
extract_meta_data_AE(ae.id, temp.dir = temp.dir)
}
\description{
Download the IDF file from ArrayExpress and extract some of the information
}
\keyword{internal}
