% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_geo.R
\name{extract_meta_data_GEO}
\alias{extract_meta_data_GEO}
\title{Extract meta data from GEO}
\usage{
extract_meta_data_GEO(gse, temp.dir = tempdir(), platform = NULL)
}
\description{
Extract platform information and information from the experimentData()
slot from the ExpressionSet of a GEO study
}
\keyword{internal}
