% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_pheno_data.R
\name{extract_pheno_data}
\alias{extract_pheno_data}
\title{Extract phenotype data}
\usage{
extract_pheno_data(study.id, temp.dir = tempdir(), platform = NULL)
}
\arguments{
\item{study.id}{[character(1)] study identifier (GSEXXX for GEO, E-MTAB-XXX
for ArrayExpress)}

\item{temp.dir}{[character(1)] directory where temporary files should be
stored}

\item{platform}{[character(1)] platform to be selected in GEO
studies which use several platforms}
}
\value{
[data.frame] phenotype data for each sample and all variables
provided by GEO or ArrayExpress
}
\description{
Extract phenotype data from GEO or ArrayExpress depending on study
identifier.
}
\details{
All phenotype data available at GEO or AE is returned without any
modifications.
}
