% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_arrayexpress.R
\name{extract_pheno_data_AE}
\alias{extract_pheno_data_AE}
\title{Extract phenotype data from ArrayExpress}
\usage{
extract_pheno_data_AE(ae.id, temp.dir = tempdir())
}
\description{
Download the SDRF file from ArrayExpress and return it as a data.frame
}
\keyword{internal}
