% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_geo.R
\name{extract_pheno_data_GEO}
\alias{extract_pheno_data_GEO}
\title{Extract phenotype data from GEO}
\usage{
extract_pheno_data_GEO(gse, temp.dir = tempdir(), platform = NULL)
}
\description{
Extract the pData() slot from the ExpressionSet of a GEO study
}
\keyword{internal}
