% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_batch.R
\name{extract_scan_date}
\alias{extract_scan_date}
\title{Extract scan date}
\usage{
extract_scan_date(
  se,
  col.sample.id = "geo_accession",
  array.type = "affy",
  temp.dir = tempdir(),
  tryFormats = c("\%Y-\%m-\%d", "\%Y/\%m/\%d", "\%m/\%d/\%Y", "\%m/\%d/\%y",
    "\%m-\%d-\%Y", "\%d-\%b-\%Y")
)
}
\arguments{
\item{se}{[\code{\link[SummarizedExperiment]{SummarizedExperiment-class}} object]
study data set}

\item{col.sample.id}{[character(1)] Column in colData() with
sample accession information (default: "geo_accession")}

\item{array.type}{[character(1)] type of array ("affy" (default), "agilent"
or "illumina")}

\item{temp.dir}{[character(1)] directory where temporary files should be
stored}

\item{tryFormats}{[character(n)] Possible format strings to try when
converting date as character to an object of class
\code{\link[base]{Date}}}
}
\value{
\code{\link[SummarizedExperiment]{SummarizedExperiment-class}}
object with scan date added as column scan.date in colData.
}
\description{
Extracts date of scanning from original array files (e.g. CEL for Affymetrix
and idat for Illumina arrays).
}
\details{
Currently only implemented for Affymetrix arrays.
}
