% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_pheno_data.R
\name{extract_subject_id}
\alias{extract_subject_id}
\title{Extract part of subject identifiers}
\usage{
extract_subject_id(id, ind.use)
}
\description{
Extract part of subject identifiers
}
\keyword{internal}
