% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_expr_data.R
\name{get_expr_data}
\alias{get_expr_data}
\title{Download and normalize raw array data}
\usage{
get_expr_data(
  study.id,
  pheno,
  type = "expr",
  array.type = "affy",
  platform = "illuminaHumanv4",
  temp.dir = tempdir()
)
}
\description{
Use specific functions for each array type.
}
\keyword{internal}
