% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_expr_data.R
\name{get_expr_data_affy}
\alias{get_expr_data_affy}
\title{Download and normalize raw array data}
\usage{
get_expr_data_affy(study.id, pheno, use.upc = FALSE, temp.dir = tempdir())
}
\description{
Use specific functions for each array type.
}
\keyword{'_at'.}
\keyword{(Affy)}
\keyword{BrainArray}
\keyword{CEL}
\keyword{Download}
\keyword{Ensembl}
\keyword{For}
\keyword{Keep}
\keyword{and}
\keyword{annotation}
\keyword{be}
\keyword{download}
\keyword{each}
\keyword{file}
\keyword{files}
\keyword{for}
\keyword{gene}
\keyword{genes}
\keyword{identifier}
\keyword{installed.}
\keyword{internal}
\keyword{normalize}
\keyword{only}
\keyword{package}
\keyword{relevant}
\keyword{remove}
\keyword{sample}
\keyword{suffix}
\keyword{the}
\keyword{using}
\keyword{which}
\keyword{will}
\keyword{with}
