% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_pheno_data.R
\name{harmonize_pheno_data}
\alias{harmonize_pheno_data}
\title{Harmonize phenotype data}
\usage{
harmonize_pheno_data(
  project,
  pheno,
  info.var,
  col.id,
  ind.use.id = NULL,
  cols.use,
  disease = NULL
)
}
\arguments{
\item{project}{[character(1)] name of project (used as prefix for all
harmonized variables)}

\item{pheno}{[data.frame] original phenotype data (e.g. as returned by
\code{\link{extract_pheno_data}})}

\item{info.var}{[list] project level information about variables that should
be harmonized (see Details)}

\item{col.id}{[character(1)] column with information about subject
identifiers}

\item{ind.use.id}{[numeric(1)] part of subject identifier that should be
kept after splitting using " ", "_" or "-"}

\item{cols.use}{[vector(n)] vector of columns used for harmonization, named
by variable names given in info.var}

\item{disease}{[character(1)] disease that should be set for all samples}
}
\value{
[data.frame] harmonized phenotype data
}
\description{
Extend original phenotype data by new variables with harmonized names and
values including sample and subject identifiers. Values of numeric variables
are converted to numeric. Values of character variables are mapped to
pre-specified values. For time variables numbers are extracted and
converted to days. If a study contains only patients of a specific
disease, this variable can be set globally.
}
\details{
info.var contains information about all variables that should be harmonized
within a project, i.e. across studies. The list needs to be named by the
names of the harmonized variables and each element contains:
\itemize{
\item type: either "character", "numeric" or "time"
\item values: list named by final value and original values that should be
mapped (can include regular expressions)
}
}
\examples{
info.var = list(
  age = list(type = "numeric"),
  sex = list(
    type = "character",
    values = list(
      female = c("^f$", "female", "woman"),
      male = c("^m$", "^male", "^man"))
  ))
}
