% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_se.R
\name{make_se_object}
\alias{make_se_object}
\title{Make SummarizedExperiment object}
\usage{
make_se_object(pheno, meta.data.l = NULL, assays, anno = NULL)
}
\arguments{
\item{pheno}{[data.frame] phenotype data}

\item{meta.data.l}{[list] meta data (as extracted by
\code{\link{extract_meta_data}})}

\item{assays}{[list] matrices of gene expression data, e.g. raw and voom
transformed counts for RNAseq (as prepared by
\code{\link{prepare_count_data}})) or raw and SCAN normalized expression
values for array data}

\item{anno}{[data.frame] gene annotation (e.g. gene identifier, chromosomal
position, gene names) with genes in rows and information in columns}
}
\value{
[SummarizedExperiment object] as generated by
\code{\link[SummarizedExperiment]{SummarizedExperiment}}
}
\description{
Combine expression data, phenotype data, gene annotation and meta data into
SummarizedExperiment object including checks on samples and genes.
}
