% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_pheno_data.R
\name{map_variable}
\alias{map_variable}
\title{Harmonize values of a character variable}
\usage{
map_variable(x, values, priority = NULL)
}
\description{
Map values of a character variable to pre-specified values. Print table of
mapping result.
}
\keyword{internal}
