% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_expr_data.R
\name{prepare_array_data}
\alias{prepare_array_data}
\title{Prepare array data}
\usage{
prepare_array_data(
  study.id,
  pheno,
  array.type = "affy",
  platform = NULL,
  temp.dir = tempdir()
)
}
\arguments{
\item{study.id}{[character(1)] study identifier (GSEXXX for GEO, E-MTAB-XXX
for ArrayExpress)}

\item{pheno}{[data.frame] phenotype data (needs to contain column with
sample identifiers (name contains "accession"))}

\item{array.type}{[character(1)] type of array ("affy" (default), "agilent"
or "illumina")}

\item{platform}{[character(1)] array platform (needs to be specified for
Illumina arrays (?))}

\item{temp.dir}{[character(1)] directory where temporary files should be
stored}
}
\value{
[list] assays (expr and upc) to be used in
\code{\link{make_se_object}}
}
\description{
Download raw data from GEO or ArrayExpress and normalize it using the SCAN
and UPC approaches (as implemented in \code{\link[SCAN.UPC]{SCANfast}}
and \code{\link[SCAN.UPC]{UPCfast}}).
}
