% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_expr_data.R
\name{prepare_count_data}
\alias{prepare_count_data}
\title{Prepare count data}
\usage{
prepare_count_data(counts)
}
\arguments{
\item{counts}{[data.frame or matrix] raw counts with genes in rows and
samples in columns}
}
\value{
[list] assays (counts, voom and voom.weights) to be used in
\code{\link{make_se_object}}
}
\description{
Remove genes with only zero counts across samples and perform voom
transformation (as implemented in \code{\link[limma]{voom}})
}
\examples{
# define temporary directory for storing file from recount
temp.dir = tempdir()

# get counts from recount
library(recount)
download_study(
  project = "SRP057087",
  type = "counts-gene",
  outdir = temp.dir)
counts = read.table(
  file.path(temp.dir, "counts_gene.tsv.gz"),
  header = TRUE,
  row.names = 29)

assays = prepare_count_data(counts)
}
