% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_expr_data.R
\name{run_norm_scan_affy}
\alias{run_norm_scan_affy}
\title{Download and normalize CEL files (Affy)}
\usage{
run_norm_scan_affy(study.id, pheno, type, temp.dir = tempdir(), file.name)
}
\description{
For each sample download and normalize CEL file using BrainArray annotation
for which relevant package will be installed. Keep only genes with Ensembl
gene identifier and remove suffix '_at'.
}
\keyword{internal}
