% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_geo.R
\name{wrapper_getGEO}
\alias{wrapper_getGEO}
\title{Wrapper of the GEOquery::getGEO() function}
\usage{
wrapper_getGEO(
  gse,
  GSEMatrix = TRUE,
  AnnotGPL = FALSE,
  getGPL = FALSE,
  platform = NULL,
  temp.dir = tempdir()
)
}
\description{
The wrapper function adds some basic validity checks including format of the
study identifier and number of different platforms used in the study.
}
\details{
This functions results in an error if the study identifier does not start
with GSE and if more than one platform is used in a study. If the latter is
the case, the desired platform can be set using the platform parameter.
}
\keyword{internal}
