% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_pheno_data.R
\name{check_subject_info}
\alias{check_subject_info}
\title{Check subject level information}
\usage{
check_subject_info(pheno, cols.subject)
}
\arguments{
\item{pheno}{[data.frame] harmonized phenotype data}

\item{cols.subject}{[vector(n)] column names to be used for check}
}
\description{
Check if subject information is consistent across all samples.
}
