% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_expr_data.R
\name{prepare_count_data}
\alias{prepare_count_data}
\title{Prepare count data}
\usage{
prepare_count_data(counts)
}
\arguments{
\item{counts}{[data.frame or matrix] raw counts with genes in rows and
samples in columns}
}
\value{
[list] assays (counts, voom and voom.weights) to be used in XXX
}
\description{
Remove genes with only zero counts across samples and perform voom
transformation (as implemented in \code{\link[limma]{voom}})
}
